+++ src/App.css
@@ -11,7 +11,7 @@
   --footer-height: 94px;
 }
 
-.App {
+.app {
   position: relative;
   min-height: 100vh;
   padding-bottom: var(--footer-height);
